requirements.txt
coverage==4.4.1
dice==1.1.1
nose==1.3.7
python-telegram-bot==8.0
